Release/0.16.0 (#131)
* feat: add support for quickbar actions (#123)

* feat: add support for quickbar actions

* chore: refactor fireDOMEvent function

* fix: correctly apply auto_padding on desktop bottom position (#121)

* Feature/127 edit mode action (#128)

* fix: TS errors in color.ts

* refactor: split config and types

* feat: new open-edit-mode action

* fix: test imports

* fix: set hass internal state as property

* fix: useDefineForClassFields=false

* feat: new `floating` mode for mobile devices

* fix: prevent adding `auto_padding` if its value in pixels is 0

* fix: prevent styles conflicting with edit-mode

* chore: bump version

* Update README.md

---------
