+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "navbar-card",
-  "version": "0.15.0",
+  "version": "0.16.0",
   "author": "Jose Álvarez Quiroga <joseluisalvquiroga@gmail.com>",
   "repository": "git@github.com:joseluis9595/lovelace-navbar-card.git",
   "license": "MIT",
