+++ src/dom-utils.ts
@@ -2,8 +2,8 @@ import {
   AutoPaddingConfig,
   DEFAULT_NAVBAR_CONFIG,
   NavbarCardConfig,
-  RippleElement,
-} from './types';
+} from './config';
+import { RippleElement } from './types';
 
 /**
  * Get a list of user defined navbar-card templates
@@ -56,6 +56,16 @@ const findHuiRoot = () => {
 };
 
 /**
+ * Forcefully open the edit mode of the Lovelace panel.
+ */
+export const forceOpenEditMode = () => {
+  const huiRoot = findHuiRoot();
+  if (!huiRoot?.shadowRoot) return;
+  // @ts-expect-error lovelace does not have "lovelace" property type
+  huiRoot.lovelace.setEditMode(true);
+};
+
+/**
  * Manually inject styles into the hui-root element to force dashboard padding.
  * This prevents overlaps with other cards in the dashboard.
  */
@@ -99,9 +109,13 @@ export const forceDashboardPadding = (options?: {
   // Desktop padding
   const desktopPaddingPx =
     options?.auto_padding?.desktop_px ??
-    DEFAULT_NAVBAR_CONFIG.layout?.auto_padding?.desktop_px;
+    DEFAULT_NAVBAR_CONFIG.layout?.auto_padding?.desktop_px ??
+    0;
 
-  if (['left', 'right'].includes(options?.desktop?.position ?? '')) {
+  if (
+    ['left', 'right'].includes(options?.desktop?.position ?? '') &&
+    desktopPaddingPx > 0
+  ) {
     cssText += `
       @media (min-width: ${desktopMinWidth}px) {
        :not(.edit-mode) > #view {
@@ -110,8 +124,9 @@ export const forceDashboardPadding = (options?: {
       }
     `;
   } else if (
-    options?.desktop?.position === 'bottom' ||
-    options?.desktop?.position === 'top'
+    (options?.desktop?.position === 'bottom' ||
+      options?.desktop?.position === 'top') &&
+    desktopPaddingPx > 0
   ) {
     cssText += `
       @media (min-width: ${desktopMinWidth}px) {
@@ -129,9 +144,11 @@ export const forceDashboardPadding = (options?: {
   // Mobile padding
   const mobilePaddingPx =
     options?.auto_padding?.mobile_px ??
-    DEFAULT_NAVBAR_CONFIG.layout?.auto_padding?.mobile_px;
+    DEFAULT_NAVBAR_CONFIG.layout?.auto_padding?.mobile_px ??
+    0;
 
-  cssText += `
+  if (mobilePaddingPx > 0) {
+    cssText += `
       @media (max-width: ${mobileMaxWidth}px) {
         :not(.edit-mode) > hui-view:after {
           content: "";
@@ -139,9 +156,10 @@ export const forceDashboardPadding = (options?: {
           height: ${mobilePaddingPx}px;
           width: 100%;
           background-color: transparent; 
+          }
         }
-      }
-    `;
+      `;
+  }
 
   // Append styles to hui-root
   if (!styleEl) {
